crawler/requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
opencc-python-reimplemented==0.1.6